requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
mongoengine==0.27.0
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
pydantic[email]==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.2.0
reportlab==4.1.0
pandas==2.2.1
openpyxl==3.1.2
httpx==0.27.0
jinja2==3.1.3
qrcode==7.4.2
